task_3/requirements.txt
absl-py==2.3.1
annotated-types==0.7.0
anyio==4.12.0
astunparse==1.6.3
beautifulsoup4==4.14.3
cachetools==6.2.2
certifi==2025.11.12
charset-normalizer==3.4.4
contourpy==1.3.3
cycler==0.12.1
Cython==3.2.2
flatbuffers==25.9.23
fonttools==4.61.0
gast==0.7.0
google-auth==2.43.0
google-genai==1.53.0
google-pasta==0.2.0
grpcio==1.76.0
h11==0.16.0
h5py==3.15.1
httpcore==1.0.9
httpx==0.28.1
idna==3.11
joblib==1.5.2
keras==3.12.0
kiwisolver==1.4.9
libclang==18.1.1
lxml==6.0.2
Markdown==3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.7
mdurl==0.1.2
ml_dtypes==0.5.4
namex==0.1.0
numpy==2.3.5
nvidia-cublas-cu12==12.9.1.4
nvidia-cuda-cupti-cu12==12.9.79
nvidia-cuda-nvcc-cu12==12.9.86
nvidia-cuda-nvrtc-cu12==12.9.86
nvidia-cuda-runtime-cu12==12.9.79
nvidia-cudnn-cu12==9.17.0.29
nvidia-cufft-cu12==11.4.1.4
nvidia-curand-cu12==10.3.10.19
nvidia-cusolver-cu12==11.7.5.82
nvidia-cusparse-cu12==12.5.10.65
nvidia-nccl-cu12==2.28.9
nvidia-nvjitlink-cu12==12.9.86
opt_einsum==3.4.0
optree==0.18.0
packaging==25.0
pandas==2.3.3
patsy==1.0.2
pillow==12.0.0
pmdarima==2.1.1
protobuf==6.33.2
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
requests==2.32.5
rich==14.2.0
rsa==4.9.1
scikit-learn==1.7.2
scipy==1.16.3
setuptools==80.9.0
six==1.17.0
soupsieve==2.8
statsmodels==0.14.6
tenacity==9.1.2
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tensorflow==2.20.0
termcolor==3.2.0
threadpoolctl==3.6.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.6.0
websockets==15.0.1
Werkzeug==3.1.4
wheel==0.45.1
wrapt==2.0.1
